Generalise in preparation for re-use
